Factor out TLA+ monitor instrumentation
